requirements-20241102.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.2
django-allauth==65.1.0
django-bootstrap5==24.3
django-countries==7.6.1
django-crispy-forms==2.3
django-filter==24.3
django-localflavor==4.0
django-requests==0.0.2
django-tables2==2.7.0
idna==3.10
neapolitan==24.7
python-stdnum==1.20
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
